% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskArchive.R
\name{archiveTask}
\alias{archiveTask}
\title{Create an archive containing the files of a given task.}
\usage{
archiveTask(task, file, overwrite = FALSE, ...)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{file}{full name of the output zip file}

\item{overwrite}{overwrite the output zip file is it exists}

\item{...}{
  Arguments passed on to \code{\link[utils:zip]{utils::zip}}
  \describe{
    \item{\code{zipfile}}{The pathname of the zip file: tilde expansion (see
    \code{\link{path.expand}}) will be performed.}
    \item{\code{files}}{A character vector of recorded filepaths to be included.}
    \item{\code{flags}}{A character string of flags to be passed to the command:
    see \sQuote{Details}.}
    \item{\code{extras}}{An optional character vector: see \sQuote{Details}.}
    \item{\code{zip}}{A character string specifying the external command to be used.}
  }}
}
\value{
the archive file name invisibly.
}
\description{
Create an archive containing the files of a given task.
}
