% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{binaryFn}
\alias{binaryFn}
\title{Get path of binary file.}
\usage{
binaryFn(task, type, ext = "rds", subdir = NULL, dirCreate = TRUE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
File path.
}
\description{
Get path of binary file.
}
