% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskFiles.R
\name{catReport}
\alias{catReport}
\title{Output R object using function cat.}
\usage{
catReport(
  x,
  task,
  type,
  ext = "txt",
  subdir = NULL,
  dirCreate = TRUE,
  sep = "\\n",
  eof = "\\n",
  ...
)
}
\arguments{
\item{x}{R object to output.}

\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}

\item{sep}{separator}

\item{eof}{EOF}

\item{...}{\R objects (see \sQuote{Details} for the types of objects
    allowed).}
}
\value{
the file name invisibly.
}
\description{
Output R object using function cat.
}
