% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTask.R
\name{createTask}
\alias{createTask}
\title{Create a task.}
\usage{
createTask(
  project,
  package,
  taskname,
  sponsor = getTaskSponsor(),
  author = getTaskAuthor()
)
}
\arguments{
\item{project}{Project name.}

\item{package}{Package name.}

\item{taskname}{Task name.}

\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}

\item{author}{Author name, system username by default.}
}
\value{
An \code{\link{D4TAlinkTask}} object
}
\description{
Create a task.
}
\keyword{internal}
