% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{D4getenv}
\alias{D4getenv}
\title{Get env var.}
\usage{
D4getenv(name, desc, quiet = FALSE)
}
\arguments{
\item{name}{name of variable}

\item{desc}{description of variable}

\item{quiet}{suppress error messages, default=FALSE.}
}
\value{
Environment variable.
}
\description{
Get env var.
}
