% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4TAlink-common-args-doc.R
\name{D4TAlink-common-args}
\alias{D4TAlink-common-args}
\title{Arguments used across the functions of the D4TAlink package.}
\arguments{
\item{project}{Project name.}

\item{package}{Package name.}

\item{taskname}{Task name.}

\item{author}{Author name, system username by default.}

\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}

\item{rootpath}{Path of the task repository, default set by \code{\link{setTaskRoot}}.}

\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{suffix}{Filename suffix, used to develop scripts for sub-analyses for a given task, default NA.}

\item{dirType}{Directory type, e.g. 'bin' or 'data' or 'doc'.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}

\item{pathgen}{optional function returning a list of paths, currently \code{\link{pathsGLPG}} or \code{\link{pathsPMS}}.}
}
\value{
No return value, used for the documentation of
the functions of the package.
}
\description{
Arguments used across the functions of the D4TAlink package.
}
