% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{datasourceFn}
\alias{datasourceFn}
\title{Get path of data source file.}
\usage{
datasourceFn(task, filename, subdir = ".", dirCreate = TRUE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{filename}{name of the input file.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
File path.
}
\description{
Get path of data source file.
}
