% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{docDir}
\alias{docDir}
\title{Get path of documentation directory.}
\usage{
docDir(task, subdir = NULL, dirCreate = TRUE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
File path.
}
\description{
Get path of documentation directory.
}
