% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{DTx}
\alias{DTx}
\title{Generic function.}
\usage{
DTx(sponsor = getTaskSponsor(), task = NULL)
}
\arguments{
\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}

\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}
}
\value{
\code{NULL}.
}
\description{
Generic function.
}
