% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskRmarkdown.R
\name{formatTaskDocx}
\alias{formatTaskDocx}
\title{Replace default task fields in 'docx' file.}
\usage{
formatTaskDocx(task, ifn)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{ifn}{input file name.}
}
\value{
the file name invisibly.
}
\description{
Replace default task fields in 'docx' file.
}
