% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskAuthor}
\alias{getTaskAuthor}
\title{Get the name of the task author.}
\usage{
getTaskAuthor(quiet = FALSE)
}
\arguments{
\item{quiet}{suppress error messages, default=FALSE.}
}
\value{
The current name of the tasks author.
}
\description{
Get the name of the task author.
}
\examples{
getTaskAuthor(quiet=TRUE)
}
