% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskEnckey}
\alias{getTaskEnckey}
\title{Get the encryption key for binary data files.}
\usage{
getTaskEnckey(ask = FALSE)
}
\arguments{
\item{ask}{query encryption key from user, default FALSE.}
}
\value{
The encryption key invisibly.
}
\description{
Get the encryption key for binary data files.
}
