% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{getTaskFilepath}
\alias{getTaskFilepath}
\title{Get the path of a file.}
\usage{
getTaskFilepath(task, type, ext, dirtype, subdir = NULL, dirCreate = TRUE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{dirtype}{task directory where file is stored, i.e., 'documentation', 'code', 'data', 'data source' or 'binary data'.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
Full path to file.
}
\description{
Get the path of a file.
}
