% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{getTaskPaths}
\alias{getTaskPaths}
\title{Get the paths of the task.}
\usage{
getTaskPaths(task)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}
}
\value{
List of task's paths.
}
\description{
Get the paths of the task.
}
