% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskRoot}
\alias{getTaskRoot}
\title{Get the root of the task repository.}
\usage{
getTaskRoot(quiet = FALSE)
}
\arguments{
\item{quiet}{suppress error messages, default=FALSE.}
}
\value{
Path to the current task root.
}
\description{
Get the root of the task repository.
}
\examples{
getTaskRoot(quiet=TRUE)
}
