% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskRscriptTemplate}
\alias{getTaskRscriptTemplate}
\title{Get the path to the R script task template.}
\usage{
getTaskRscriptTemplate()
}
\value{
The path to the R script task template.
}
\description{
Get the path to the R script task template.
}
