% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskSponsor}
\alias{getTaskSponsor}
\title{Get the name of the task sponsor.}
\usage{
getTaskSponsor(quiet = FALSE)
}
\arguments{
\item{quiet}{suppress error messages, default=FALSE.}
}
\value{
The current name of the tasks sponsor.
}
\description{
Get the name of the task sponsor.
}
\examples{
getTaskSponsor(quiet=TRUE)
}
