% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskRmarkdown.R
\name{initTaskRmd}
\alias{initTaskRmd}
\title{Create task template in Rmd format.}
\usage{
initTaskRmd(task, encoding = "unknown", overwrite = FALSE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{encoding}{encoding to be assumed for input strings.  It is
    used to mark character strings as known to be in
    Latin-1 or UTF-8: it is not used to re-encode the input.  To do the
    latter, specify the encoding as part of the connection \code{con} or
    via \code{\link[base]{options}(encoding=)}: see the examples.
    See also \sQuote{Details}.
  }

\item{overwrite}{overwrite Rmd file if exists, default FALSE}
}
\value{
the file name invisibly.
}
\description{
Create task template in Rmd format.
}
