% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskRscript.R
\name{initTaskRscript}
\alias{initTaskRscript}
\title{Create task R script.}
\usage{
initTaskRscript(task, overwrite = FALSE, encoding = "unknown", suffix = NA)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{overwrite}{overwrite R file if exists, default FALSE}

\item{encoding}{encoding to be assumed for input strings.  It is
    used to mark character strings as known to be in
    Latin-1, UTF-8 or to be bytes: it is not used to re-encode the input.
    To do the
    latter, specify the encoding as part of the connection \code{con} or
    via \code{\link[base]{options}(encoding=)}: see the examples
    and \sQuote{Details}.
  }

\item{suffix}{Filename suffix, used to develop scripts for sub-analyses for a given task, default NA.}
}
\value{
the file name invisibly.
}
\description{
Create task R script.
}
