% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTasks.R
\name{listTasks}
\alias{listTasks}
\title{List details of all tasks stored in the root directory.}
\usage{
listTasks(
  project = NULL,
  package = NULL,
  sponsor = NULL,
  rootpath = getTaskRoot()
)
}
\arguments{
\item{project}{Project name.}

\item{package}{Package name.}

\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}

\item{rootpath}{Path of the task repository, default set by \code{\link{setTaskRoot}}.}
}
\value{
\code{\link{data.frame}} with the following information for tasks "sponsor", "project", "package", "task".
}
\description{
List details of all tasks stored in the root directory.
}
