% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTask.R
\name{loadTask}
\alias{loadTask}
\title{Load a task.}
\usage{
loadTask(
  project,
  package,
  taskname,
  sponsor = getTaskSponsor(),
  author = getTaskAuthor(),
  quiet = FALSE
)
}
\arguments{
\item{project}{Project name.}

\item{package}{Package name.}

\item{taskname}{Task name.}

\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}

\item{author}{Author name, system username by default.}

\item{quiet}{issue warning if file does not exists.}
}
\value{
Object of class \code{\link{D4TAlinkTask}} or NULL if the \code{task} does not exists.
}
\description{
Load a task.
}
