% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathGenerators.R
\name{pathsPMS}
\alias{pathsPMS}
\title{Task paths generator.}
\usage{
pathsPMS(project, package, taskname, sponsor)
}
\arguments{
\item{project}{Project name.}

\item{package}{Package name.}

\item{taskname}{Task name.}

\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}
}
\value{
a list of file paths
}
\description{
The paths are:
datasrc: [ROOT]/[sponsor]/PMS_data/[project]/[package]/datasource
data:    [ROOT]/[sponsor]/PMS_data/[project]/[package]/[taskname]
bin:     [ROOT]/[sponsor]/PMS_data/[project]/[package]/[taskname]/bin
code:    [ROOT]/[sponsor]/PMS_code/[project]/[package]/R
doc:     [ROOT]/[sponsor]/PMS_documentation/[project]/[package]/[taskname]
log:     [ROOT]/[sponsor]/PMS_data/[project]/[package]/[taskname]/log
}
