% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskPaths.R
\name{pngReportFn}
\alias{pngReportFn}
\title{Get path of \code{png} output file.}
\usage{
pngReportFn(task, type, ext = "png", subdir = NULL)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{subdir}{(optional) Subdirectory.}
}
\value{
File path.
}
\description{
Get path of \code{png} output file.
}
