% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskFiles.R
\name{readBinary}
\alias{readBinary}
\title{Restore R object from binary file.}
\usage{
readBinary(
  task,
  type,
  subdir = NULL,
  dirCreate = FALSE,
  ask = FALSE,
  quiet = FALSE,
  password = NULL
)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}

\item{ask}{query encryption key from user, default FALSE.}

\item{quiet}{issue warning if file does not exists.}

\item{password}{encryption password, default NULL}
}
\value{
Object stored in binary file, or \code{NULL} if file does not exist.
}
\description{
Restore R object from binary file.
}
