% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonio.R
\name{readFeather}
\alias{readFeather}
\title{Restore R data.frame from Apache Arrow feather file.}
\usage{
readFeather(task, type, subdir = NULL, dirCreate = FALSE, quiet = FALSE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}

\item{quiet}{issue warning if file does not exists.}
}
\value{
Object stored in Apache Arrow feather file, or \code{NULL} if file does not exist.
}
\description{
Restore R data.frame from Apache Arrow feather file.
}
