% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskFiles.R
\name{readReportJSON}
\alias{readReportJSON}
\title{Read JSON data into R object.}
\usage{
readReportJSON(task, type, ext = "json", subdir = NULL, dirCreate = FALSE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
the data read, or \code{NULL} if the file does not exist.
}
\description{
Read JSON data into R object.
}
