% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskRmarkdown.R
\name{rmdFn}
\alias{rmdFn}
\title{Get path of R script file name.}
\usage{
rmdFn(task)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}
}
\value{
File path.
}
\description{
Get path of R script file name.
}
