% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskFiles.R
\name{saveBinary}
\alias{saveBinary}
\title{Save R object in binary file.}
\usage{
saveBinary(
  object,
  task,
  type,
  subdir = NULL,
  dirCreate = TRUE,
  encrypt = FALSE,
  ask = FALSE
)
}
\arguments{
\item{object}{R object to serialize.}

\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}

\item{encrypt}{encrypt the output, default: FALSE. If character string, then use the string as password.}

\item{ask}{query encryption key from user, default FALSE.}
}
\value{
the file name invisibly.
}
\description{
Save R object in binary file.
}
