% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonio.R
\name{saveFeather}
\alias{saveFeather}
\title{Save R object in Apache Arrow feather file.}
\usage{
saveFeather(object, task, type, subdir = NULL, dirCreate = TRUE)
}
\arguments{
\item{object}{data.frame to serialize.}

\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
the file name invisibly.
}
\description{
Save R object in Apache Arrow feather file.
}
