% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskFiles.R
\name{saveReportJSON}
\alias{saveReportJSON}
\title{Output R object in JSON format.}
\usage{
saveReportJSON(x, task, type, ext = "json", subdir = NULL, dirCreate = TRUE)
}
\arguments{
\item{x}{R object to output.}

\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{ext}{Filename extension.}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
the file name invisibly.
}
\description{
Output R object in JSON format.
}
