% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskAuthor}
\alias{setTaskAuthor}
\title{Set the name of the tasks author.}
\usage{
setTaskAuthor(author)
}
\arguments{
\item{author}{Author name, system username by default.}
}
\value{
The current name of the tasks author.
}
\description{
Set the name of the tasks author.
}
\examples{
setTaskAuthor("Doe Johns")
}
