% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskEnckey}
\alias{setTaskEnckey}
\title{Set the encryption key for binary data files.}
\usage{
setTaskEnckey(key)
}
\arguments{
\item{key}{encryption key, if NULL then query key from user.}
}
\value{
\code{NULL} invisibly.
}
\description{
Set the encryption key for binary data files.
}
