% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskRmdTemplate}
\alias{setTaskRmdTemplate}
\title{Set the path to the Rmd task template.}
\usage{
setTaskRmdTemplate(file)
}
\arguments{
\item{file}{path to the Rmd task template.}
}
\value{
The path to the Rmd task template invisibly.
}
\description{
Set the path to the Rmd task template.
}
