% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskRoot}
\alias{setTaskRoot}
\title{Set the root of the task repository.}
\usage{
setTaskRoot(rootpath, dirCreate = FALSE)
}
\arguments{
\item{rootpath}{Path of the task repository, default set by \code{\link{setTaskRoot}}.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
Path to the current task root.
}
\description{
Set the root of the task repository.
}
