% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskRscriptTemplate}
\alias{setTaskRscriptTemplate}
\title{Set the path to the R script task template.}
\usage{
setTaskRscriptTemplate(file)
}
\arguments{
\item{file}{path to the Rmd task template.}
}
\value{
The path to the Rmd task template invisibly.
}
\description{
Set the path to the R script task template.
}
