% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskSponsor}
\alias{setTaskSponsor}
\title{Set the name of the tasks sponsor.}
\usage{
setTaskSponsor(sponsor)
}
\arguments{
\item{sponsor}{Sponsor name, default set by \code{\link{setTaskSponsor}}.}
}
\value{
The current name of the tasks sponsor.
}
\description{
Set the name of the tasks sponsor.
}
\examples{
setTaskSponsor("SQU4RE")
}
