% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskID.R
\name{taskID}
\alias{taskID}
\title{Get task identifier string.}
\usage{
taskID(task, sep = "/")
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{sep}{the field separator character, default: "/".}
}
\value{
String with task ID as:[sponsor][sep][project][sep][package][sep][task]
}
\description{
Get task identifier string.
}
