% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskRmdTemplate}
\alias{getTaskRmdTemplate}
\title{Get the path to the Rmd task template.}
\usage{
getTaskRmdTemplate(quiet = FALSE)
}
\arguments{
\item{quiet}{suppress error messages, default=FALSE.}
}
\value{
The path to the Rmd task template.
}
\description{
Get the path to the Rmd task template.
}
