% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{getTaskStructure}
\alias{getTaskStructure}
\title{Get repository directory structure.}
\usage{
getTaskStructure()
}
\value{
The directory structure function.
}
\description{
Get repository directory structure.
}
