% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskFiles.R
\name{readBinary}
\alias{readBinary}
\title{Restore R object from binary file.}
\usage{
readBinary(task, type, subdir = NULL, dirCreate = FALSE)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{type}{Filename type. If the type is an array, the cocatenation of the elements is used with separator"-". Filenames have the form [task name]_[type].[ext]}

\item{subdir}{(optional) Subdirectory.}

\item{dirCreate}{Logical, if TRUE (by default) the directory is created.}
}
\value{
Object stored in binary file, or \code{NULL} if file does not exist.
}
\description{
Restore R object from binary file.
}
