% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskRmarkdown.R
\name{rmdFn}
\alias{rmdFn}
\title{Get path of R script file name.}
\usage{
rmdFn(task, suffix = NA)
}
\arguments{
\item{task}{Object of class \code{\link{D4TAlinkTask}}, as created by \code{\link{initTask}}.}

\item{suffix}{Filename suffix, used to develop scripts for sub-analyses for a given task, default NA.}
}
\value{
File path.
}
\description{
Get path of R script file name.
}
