% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskGlobals.R
\name{setTaskRmdTemplate}
\alias{setTaskRmdTemplate}
\title{Set the path to the Rmd task template.}
\usage{
setTaskRmdTemplate(file, encoding = "unknown")
}
\arguments{
\item{file}{path to the Rmd task template.}

\item{encoding}{encoding to be assumed for input strings.  It is
    used to mark character strings as known to be in
    Latin-1, UTF-8 or to be bytes: it is not used to re-encode the input.
    To do the
    latter, specify the encoding as part of the connection \code{con} or
    via \code{\link[base]{options}(encoding=)}: see the examples
    and \sQuote{Details}.
  }
}
\value{
The path to the Rmd task template invisibly.
}
\description{
Set the path to the Rmd task template.
}
